requirements.txt
anyio==3.6.2
arxiv==1.4.2
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
feedparser==6.0.10
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
mypy-extensions==0.4.3
notion-client==2.0.0
numpy==1.24.1
pathspec==0.10.3
platformdirs==2.6.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.13
